requirements.txt
matplotlib==3.4.3
numpy==1.20.3
PyQt5==5.15.9
PyQt5_sip==12.12.1
pyserial==3.5
pylablib==1.4.3